# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vehicles": {
      "post": {
        "operationId": "VehiclesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new visiting vehicle record",
        "tags": [
          "Visiting Vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation start date (created_at >= startDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation end date (created_at <= endDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryGateId",
            "required": false,
            "in": "query",
            "description": "Filter by entry gate ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exitGateId",
            "required": false,
            "in": "query",
            "description": "Filter by exit gate ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated visiting vehicles"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all visiting vehicles with pagination, search, sorting, and filters",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/export": {
      "get": {
        "operationId": "VehiclesController_exportToExcel_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation start date (created_at >= startDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation end date (created_at <= endDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryGateId",
            "required": false,
            "in": "query",
            "description": "Filter by entry gate ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exitGateId",
            "required": false,
            "in": "query",
            "description": "Filter by exit gate ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return exported visiting vehicles as Excel"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export visiting vehicles to Excel with pagination, search, sorting, and filters",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/active": {
      "get": {
        "operationId": "VehiclesController_findAllActive_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "gateId",
            "required": false,
            "in": "query",
            "description": "Filter by entry gate ID",
            "schema": {
              "example": "c7ce6251-4e95-48ea-b697-e1cacf5d2ad4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated active visiting vehicles"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all actively entered visiting vehicles with pagination",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return visiting vehicle"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get visiting vehicle by ID",
        "tags": [
          "Visiting Vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a visiting vehicle record",
        "tags": [
          "Visiting Vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deleted successfully"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a visiting vehicle record",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/entry": {
      "post": {
        "operationId": "VehiclesController_registerEntry_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVehicleEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle entry registered successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register a visiting vehicle entry",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}/exit": {
      "post": {
        "operationId": "VehiclesController_registerExit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVehicleExitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle exit registered successfully"
          },
          "400": {
            "description": "Vehicle is not in ENTERED status"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register a visiting vehicle exit",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{id}/cancel": {
      "post": {
        "operationId": "VehiclesController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle cancelled successfully"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel a visiting vehicle record",
        "tags": [
          "Visiting Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Terminal Port Management System (TPMS) API",
    "description": "API for managing and monitoring vehicles and visitors entering/exiting a terminal port, with JWT authentication and RBAC",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "example": "YGN-5678"
          },
          "vehicleType": {
            "type": "string",
            "example": "Sedan"
          },
          "vehicleModel": {
            "type": "string",
            "example": "Toyota Corolla"
          },
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrLicense": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Delivery"
          },
          "entryGateId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "plateNumber",
          "visitorName"
        ]
      },
      "RegisterVehicleEntryDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "example": "YGN-5678"
          },
          "vehicleType": {
            "type": "string",
            "example": "Sedan"
          },
          "vehicleModel": {
            "type": "string",
            "example": "Toyota Corolla"
          },
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrLicense": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Delivery"
          },
          "entryGateId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "plateNumber",
          "visitorName",
          "entryGateId"
        ]
      },
      "RegisterVehicleExitDto": {
        "type": "object",
        "properties": {
          "exitGateId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "exitGateId"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "example": "YGN-5678"
          },
          "vehicleType": {
            "type": "string",
            "example": "Sedan"
          },
          "vehicleModel": {
            "type": "string",
            "example": "Toyota Corolla"
          },
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrLicense": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Delivery"
          },
          "entryGateId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENTERED",
              "EXITED",
              "CANCELLED"
            ]
          },
          "exitGateId": {
            "type": "string"
          }
        }
      }
    }
  }
}